remove redundant API calls tests
